Create Yu_Jin_2019.yml
The exact roles of these genes are not investigated in this paper. The authors used bioinformatic methods to propose genes associated with germination but have not studied what the genes do. Is this a confidence level of 4?
Because the gene's functions were not discussed in detail, I included the ontologies from their Legume Info entries, but I don't have a citation for this information. Is that okay?