Create Wang_Guo_2021.yml
The Yang Lan paper refers to STF4 as well as STF3, but there is no record of its existence in Soybase or GenBank.  Most papers treat them as a unit (STF3/4) with the same functions, so I have included STF3 but excluded STF4.  Should I make a separate entry for STF4?