Ge_Yu_2016.yml Reviewed and Corrected on 8 December 2023 (#28)
* Corrected the formats of gene_model_pub_name and gene_model_full_id (single gene name for each of these fields)

* Create phavu_Davila-Delgado_Flores-Canul_2023

SYMRK gene in Phaseolus vulgaris

* Create vigra_Liu_Zhang_2022

This is a Vigna radiata gene which was studied in transgenic Arabidopsis.  Many of its suspected functions have not been experimentally validated.

* Commit for its own sake

* November 17 2023 Cai_Wang_2017.yml

* November 22 2023 Cai_Wang_2017.yml

* Remove some mis-placed files

* December 1 Commit

* Create Wang_Yuan_2019.yml

This is the only paper I found discussing the role LCL genes in maintaining circadian rhythm, and it doesn't tease apart the roles of the four LCL genes discussed. 
The authors tested the function of LCLb2 with transgenic expression and knockouts. They knocked out LCLa1, LCLa2, LCLb1, and LCLb2 all at once, so there is no way to distinguish them from each other.
Therefore, I have not included entries of LCLa1, LCLa2, and LCLb1.

* Updated Chen_Wang_2016.yml from skalberer

* Ge_Yu_2016.yml Reviewed and Corrected on 7 December 2023

* Ge_Yu_2016.yml Reviewed and Corrected on 8 December 2023

* Indrasumunar_Searle Change Year in Name

---------

Co-authored-by: StevenCannon-USDA <steven.cannon@ars.usda.gov>
Co-authored-by: MarleneDH <136021077+MarleneDH@users.noreply.github.com>